refactor(models): unify pay and expense models for consistency and clarity
- Standardized PayIncome, PayExpenseBuy, and PayExpenseOther models
  with consistent JSON tags and validation rules.
- Replaced 'credit card' with 'credit_card' in payment method validation.
- Added new schema for expenses (internal/schemas/expense.go) and
  removed obsolete schemas (expense_buy.go, expense_other.go).
- Updated related controllers, repositories, and schemas to align with
  new data structure.
- Improved naming consistency across cash register and income modules.

+++ internal/models/cash_register.go
@@ -3,15 +3,15 @@ package models
 import "time"
 
 type CashRegister struct {
-	ID           uint      `gorm:"primaryKey;autoIncrement" json:"id"`
-	PointSaleID  uint      `gorm:"not null" json:"point_sale_id"`
+	ID           int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+	PointSaleID  int64     `gorm:"not null" json:"point_sale_id"`
 	PointSale    PointSale `gorm:"foreignKey:PointSaleID;references:ID" json:"point_sale"`
-	MemberOpenID uint      `gorm:"not null" json:"user_open_id"`
+	MemberOpenID int64     `gorm:"not null" json:"user_open_id"`
 	MemberOpen   Member    `gorm:"foreignKey:MemberOpenID;references:ID" json:"member_open"`
 	OpenAmount   float64   `gorm:"" json:"open_amount"`
 	HourOpen     time.Time `gorm:"" json:"hour_open"`
 
-	MemberCloseID *uint      `json:"user_close_id"`
+	MemberCloseID *int64     `json:"user_close_id"`
 	MemberClose   *Member    `gorm:"foreignKey:MemberCloseID;references:ID" json:"member_close"`
 	CloseAmount   *float64   `gorm:"" json:"close_amount"`
 	HourClose     *time.Time `gorm:"" json:"hour_close"`
